large installer updates
'longification'


git-svn-id: file:///source/coreland/corelib/trunk@581 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ fmt.h
@@ -28,7 +28,7 @@ unsigned int fmt_ulonglongo(char *, unsigned long long);
 unsigned int fmt_float(char *, float *);
 unsigned int fmt_double(char *, double *);
 
-unsigned int fmt_str(char *, const char *);
-unsigned int fmt_nstr(char *, const char *, unsigned int);
+unsigned long fmt_str(char *, const char *);
+unsigned long fmt_nstr(char *, const char *, unsigned long);
 
 #endif
